Split Product into 4 smaller modules
